Added Credit Renewal

+++ src/main/java/com/srichell/microservices/ratelimit/app/config/RateLimitConfig.java
@@ -41,7 +41,7 @@ public class RateLimitConfig {
         return requestsPerMinute;
     }
 
-    public RateLimitConfig setRequestsPerMinute(int requestsPerMinute) {
+    public RateLimitConfig setRequestsPerMinute(Long requestsPerMinute) {
         this.requestsPerMinute = requestsPerMinute;
         return this;
     }
